Refactoring spider, process.
