+++ api/context/ingesting/src/Infrastructure/Framework/services.yaml
@@ -10,25 +10,27 @@ services:
     _defaults:
         autowire: true      # Automatically injects dependencies in your services.
         autoconfigure: true # Automatically registers your services as commands, event subscribers, etc.
+        bind:
+            '$ecotoneCommandRunner': "@console.ecotone:run"
 
     # makes classes in src/ available to be used as services
     # this creates a service per class whose id is the fully-qualified class name
-    App\:
-        resource: '../src/'
-        exclude:
-            - '../src/DependencyInjection/'
-            - '../src/Entity/'
-            - '../src/Kernel.php'
+    Ingesting\:
+        resource: '../../../src/'
+#        exclude:
+#            - '../src/DependencyInjection/'
+#            - '../src/Entity/'
+#            - '../src/Kernel.php'
 
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
 
-    Enqueue\Dbal\DbalConnectionFactory:
-        class: Enqueue\Dbal\DbalConnectionFactory
-        factory: ['Ecotone\Dbal\DbalConnection', 'fromDsn']
-        arguments: ['%env(resolve:DATABASE_URL)%']
-
-    Enqueue\AmqpExt\AmqpConnectionFactory:
-        class: Enqueue\AmqpExt\AmqpConnectionFactory
-        arguments:
-            - '%env(resolve:ENQUEUE_DSN)%'
+#    Enqueue\Dbal\DbalConnectionFactory:
+#        class: Enqueue\Dbal\DbalConnectionFactory
+#        factory: ['Ecotone\Dbal\DbalConnection', 'fromDsn']
+#        arguments: ['%env(resolve:DATABASE_URL)%']
+#
+#    Enqueue\AmqpExt\AmqpConnectionFactory:
+#        class: Enqueue\AmqpExt\AmqpConnectionFactory
+#        arguments:
+#            - '%env(resolve:ENQUEUE_DSN)%'
