+++ api/context/ingesting/src/Infrastructure/Spider/ItemProcessor/ForkedScoreProcessorItemProcessor.php
@@ -1,22 +1,21 @@
 <?php declare(strict_types=1);
 
-namespace App\Spider\ItemProcessor;
+namespace Ingesting\Infrastructure\Spider\ItemProcessor;
 
 use RoachPHP\ItemPipeline\ItemInterface;
 use RoachPHP\ItemPipeline\Processors\ItemProcessorInterface;
 use RoachPHP\Support\Configurable;
 
-final class StarredScoreProcessorItemProcessor implements ItemProcessorInterface
+final class ForkedScoreProcessorItemProcessor implements ItemProcessorInterface
 {
     use Configurable;
 
     public function processItem(ItemInterface $item): ItemInterface
     {
-        $minStarredScore = (int) $this->option('min_starred_score');
-
-        if ($minStarredScore >= (int) $item->get('starred', 0)) {
+        $minForkScore = (int) $this->option('min_fork_score');
+        if ($minForkScore >= (int) $item->get('fork', 0)) {
             return $item->drop(
-                \sprintf('Fewer than %s starred scored', $this->option('min_starred_score'))
+                \sprintf('Fewer than %s fork scored', $this->option('min_fork_score'))
             );
         }
 
@@ -26,7 +25,7 @@ final class StarredScoreProcessorItemProcessor implements ItemProcessorInterface
     private function defaultOptions(): array
     {
         return [
-            'min_starred_score' => 1,
+            'min_fork_score' => 1,
         ];
     }
 }
