+++ api/context/ingesting/src/Infrastructure/Spider/ItemProcessor/StarredScoreProcessorItemProcessor.php
@@ -1,21 +1,22 @@
 <?php declare(strict_types=1);
 
-namespace App\Spider\ItemProcessor;
+namespace Ingesting\Infrastructure\Spider\ItemProcessor;
 
 use RoachPHP\ItemPipeline\ItemInterface;
 use RoachPHP\ItemPipeline\Processors\ItemProcessorInterface;
 use RoachPHP\Support\Configurable;
 
-final class ForkedScoreProcessorItemProcessor implements ItemProcessorInterface
+final class StarredScoreProcessorItemProcessor implements ItemProcessorInterface
 {
     use Configurable;
 
     public function processItem(ItemInterface $item): ItemInterface
     {
-        $minForkScore = (int) $this->option('min_fork_score');
-        if ($minForkScore >= (int) $item->get('fork', 0)) {
+        $minStarredScore = (int) $this->option('min_starred_score');
+
+        if ($minStarredScore >= (int) $item->get('starred', 0)) {
             return $item->drop(
-                \sprintf('Fewer than %s fork scored', $this->option('min_fork_score'))
+                \sprintf('Fewer than %s starred scored', $this->option('min_starred_score'))
             );
         }
 
@@ -25,7 +26,7 @@ final class ForkedScoreProcessorItemProcessor implements ItemProcessorInterface
     private function defaultOptions(): array
     {
         return [
-            'min_fork_score' => 1,
+            'min_starred_score' => 1,
         ];
     }
 }
